projects/smart-pdf-rotation/requirements.txt
pillow
pytesseract
opencv-python
imutils
numpy
